js -> ts, prisma for db
